cleanup

+++ Chapter_09_Pi4J/java-pi4j-spi/src/main/java/be/webtechie/pi4jgpio/characters/AsciiCharacter.java
@@ -97,6 +97,6 @@ public enum AsciiCharacter {
     }
 
     public static AsciiCharacter getByChar(char character) {
-        return getByAscii((int) character);
+        return getByAscii(character);
     }
 }
